Evolución del sistema de plugins
- Se desarrolla el modelo para permitir que los plugins puedan agregar un nuevo menú a la interfaz de la aplicación.
- Se desarrolla el modelo para permitir que los plugins puedan agregar nuevos comandos para ser utilizados desde línea de comandos al utilizar la aplicación.
- Se crea un sistema de permisos que restringe los modos de integración de los plugins. Si un plugin no declara un permiso, no podrá integrarse del modo que permite el permiso (instalación del plugin, preprocesado de la firma, agregar botones a la interfaz, etc.).
- Se crea un permiso para solicitar el reinicio de la aplicación tras la instalación y desinstalación de un plugin.
- Se externaliza toda la funcionalidad de cálculo y comprobación de hashes a un nuevo plugin (afirma-ui-simple-plugin-hash).
- Se agregan los permisos necesarios al plugin de validación de certificados.

+++ afirma-simple/src/main/java/es/gob/afirma/standalone/plugins/PluginsManager.java
@@ -258,22 +258,27 @@ public class PluginsManager {
 			throw new PluginException("No se pudo cargar el plugin recien importado", e); //$NON-NLS-1$
 		}
 
-		// Ejecutamos la operacion de instalacion del propio plugin
-		try {
-			plugin.install();
-		}
-		catch (final Exception e) {
-			LOGGER.log(Level.SEVERE, "Ocurrio un error al instalar el plugin. Lo desinstalamos", e); //$NON-NLS-1$
+		final PluginInfo info = plugin.getInfo();
+
+		// Si el plugin tenia el permiso de instalacion, se ejecuta su operacion
+		// de instalacion
+		if (PermissionChecker.check(info, Permission.INSTALL)) {
 			try {
-				uninstallPlugin(plugin);
+				plugin.install();
 			}
-			catch (final Exception ex) {
-				LOGGER.warning("No se han podido eliminar los ficheros importados del plugin"); //$NON-NLS-1$
+			catch (final Exception e) {
+				LOGGER.log(Level.SEVERE, "Ocurrio un error al instalar el plugin. Lo desinstalamos", e); //$NON-NLS-1$
+				try {
+					uninstallPlugin(plugin);
+				}
+				catch (final Exception ex) {
+					LOGGER.warning("No se han podido eliminar los ficheros importados del plugin"); //$NON-NLS-1$
+				}
+				if (e instanceof PluginControlledException) {
+					throw (PluginControlledException) e;
+				}
+				throw new PluginException("Ocurrio un error al instalar el plugin", e); //$NON-NLS-1$
 			}
-			if (e instanceof PluginControlledException) {
-				throw (PluginControlledException) e;
-			}
-			throw new PluginException("Ocurrio un error al instalar el plugin", e); //$NON-NLS-1$
 		}
 
 		// Agregamos el plugin a la lista de plugins instalados
@@ -436,15 +441,19 @@ public class PluginsManager {
 	 */
 	public void uninstallPlugin(final AfirmaPlugin plugin) throws IOException {
 
-		// Ejecutamos el proceso de desinstalacion del propio plugin
-		try {
-			plugin.uninstall();
-		} catch (final Exception e) {
-			LOGGER.log(Level.WARNING, "El proceso de desinstalacion interno del plugin devolvio un error", e); //$NON-NLS-1$
+		final PluginInfo info = plugin.getInfo();
+
+		// Ejecutamos el proceso de desinstalacion del plugin si este lo tenia declarado
+		if (PermissionChecker.check(info, Permission.UNINSTALL)) {
+			try {
+				plugin.uninstall();
+			} catch (final Exception e) {
+				LOGGER.log(Level.WARNING, "El proceso de desinstalacion interno del plugin devolvio un error", e); //$NON-NLS-1$
+			}
 		}
 
 		// Descargamos el plugin de memoria
-		final String internalName = plugin.getInfo().getInternalName();
+		final String internalName = info.getInternalName();
 		this.pluginsLoadedList.remove(plugin);
 		((URLClassLoader) plugin.getClass().getClassLoader()).close();
